A* implemented

+++ A2-2x2 Solver/RubiksSolver/src/rubikssolver/Node.java
@@ -2,12 +2,12 @@ package rubikssolver;
 
 public class Node implements Comparable<Node>{
     public Cube state;
-    public int priority;
+    public double priority;
     public Node parent;
     public MOVES lastMove;
     public int pathCost;
 
-    public Node(Cube state, int priority, Node parent, MOVES lastMove, int pathCost) {
+    public Node(Cube state, double priority, Node parent, MOVES lastMove, int pathCost) {
         this.state = state;
         this.priority = priority;
         this.parent = parent;
@@ -17,6 +17,6 @@ public class Node implements Comparable<Node>{
 
     @Override
     public int compareTo(Node other) {
-        return Integer.compare(this.priority, other.priority);
+        return Double.compare(this.priority, other.priority);
     }
 }
